server/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pgvector==0.3.6

# Redis
redis[hiredis]==5.2.1

# Celery
celery[redis]==5.4.0

# Auth
pyjwt==2.10.1
bcrypt==4.2.1
pyotp==2.9.0

# Validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# File parsing
pypdf2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2

# LLM clients
openai==1.58.1
httpx==0.28.1

# Utils
python-jose[cryptography]==3.3.0
pyahocorasick==2.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
